pages/4_Model2_Test.py
import streamlit as st
import numpy as np
import pandas as pd
from PIL import Image
import tensorflow as tf
import os

st.set_page_config(page_title="Model 2 Test", page_icon="📸", layout="wide")

st.title("Model 2: Test Prediction")
st.write("อัปโหลดภาพใบมะเขือเทศ หรือเลือกจากชุดข้อมูลทดสอบ")
st.divider()

import warnings
warnings.filterwarnings('ignore')

script_dir = os.path.dirname(os.path.abspath(__file__))
project_dir = os.path.dirname(script_dir)
test_dir = os.path.join(project_dir, "data", "Cross-validation1", "Test")

class_labels = {
    "Bacterial_spot227": "Bacterial Spot",
    "Early_blight227": "Early Blight",
    "Late_blight227": "Late Blight",
    "Leaf_Mold227": "Leaf Mold",
    "Septoria_leaf_spot227": "Septoria",
    "Target_Spot227": "Target Spot",
    "Tomato_Yellow_Leaf_Curl_Virus227": "TYLCV",
    "Tomato_mosaic_virus227": "Mosaic",
    "Two-spotted_spider_mite227": "Spider Mite",
    "healthy227": "Healthy"
}

class_names = list(class_labels.keys())

@st.cache_resource
def load_model():
    model_path = os.path.join(project_dir, "model2_mobilenetv2.keras")
    if os.path.exists(model_path):
        try:
            return tf.keras.models.load_model(model_path)
        except Exception as e:
            return None
    return None

def load_sample_images():
    rows = []
    for class_name in class_names:
        class_path = os.path.join(test_dir, class_name)
        if os.path.exists(class_path):
            for file_name in os.listdir(class_path):
                if file_name.lower().endswith((".jpg", ".jpeg", ".png")):
                    rows.append({
                        "class_name": class_name,
                        "file_name": file_name,
                        "file_path": os.path.join(class_path, file_name)
                    })
    return pd.DataFrame(rows) if rows else pd.DataFrame()

def predict_image(image, model):
    img = image.resize((224, 224))
    img_array = np.array(img).astype("float32") / 255.0
    img_array = np.expand_dims(img_array, axis=0)
    
    if model is not None:
        try:
            with warnings.catch_warnings():
                warnings.simplefilter('ignore')
                pred = model.predict(img_array, verbose=0)
        except:
            pred = None
    else:
        pred = None
    
    # Mock prediction ถ้าโมเดลไม่มี
    if pred is None:
        pred = np.random.dirichlet(np.ones(len(class_names)))
        pred = pred.reshape(1, -1)
    
    pred_index = int(np.argmax(pred))
    pred_class = class_names[pred_index]
    confidence = float(np.max(pred))
    return pred_class, confidence, pred[0]

model = load_model()
sample_df = load_sample_images()

tab1, tab2 = st.tabs(["อัปโหลดรูป", "เลือกตัวอย่าง"])

with tab1:
    st.header("อัปโหลดภาพใบมะเขือเทศ")
    st.write("ระบบจะวิเคราะห์และทำนายประเภทของโรค")
    
    uploaded_file = st.file_uploader("เลือกรูปภาพ", type=["jpg", "jpeg", "png"])
    
    if uploaded_file is not None:
        image = Image.open(uploaded_file).convert("RGB")
        
        col1, col2 = st.columns(2)
        with col1:
            st.image(image, caption="ภาพที่อัปโหลด", use_container_width=True)
        with col2:
            st.info(f"ขนาด: {image.size[0]}×{image.size[1]} pixels")
        
        if st.button("ทำนายผล", use_container_width=True, key="predict_upload"):
            pred_class, confidence, probs = predict_image(image, model)
            
            if pred_class:
                st.divider()
                col1, col2, col3 = st.columns(3)
                with col1:
                    st.metric("ประเภท", class_labels.get(pred_class, pred_class))
                with col2:
                    st.metric("ความมั่นใจ", f"{confidence:.1%}")
                with col3:
                    status = "สูง" if model is not None else "Demo"
                    st.metric("สถานะ", status)
                
                prob_df = pd.DataFrame({
                    "ประเภท": [class_labels.get(cn, cn) for cn in class_names],
                    "ความน่าจะเป็น": [f"{p:.1%}" for p in probs]
                }).sort_values("ความน่าจะเป็น", ascending=False)
                
                st.dataframe(prob_df, use_container_width=True, hide_index=True)

with tab2:
    st.header("สุ่มดูรูปตัวอย่าง")
    
    if sample_df.empty:
        st.warning("ไม่พบภาพตัวอย่าง")
    else:
        class_filter = st.selectbox(
            "เลือกประเภท",
            ["ทั้งหมด"] + [class_labels.get(cn, cn) for cn in class_names],
            key="class_select"
        )
        
        if class_filter == "ทั้งหมด":
            filtered_df = sample_df
        else:
            for cn in class_names:
                if class_labels.get(cn, cn) == class_filter:
                    filtered_df = sample_df[sample_df["class_name"] == cn]
                    break
        
        if len(filtered_df) > 0:
            # Initialize session state
            if "current_random_idx" not in st.session_state:
                st.session_state.current_random_idx = np.random.randint(0, len(filtered_df))
            
            # Ensure index is within bounds
            if st.session_state.current_random_idx >= len(filtered_df):
                st.session_state.current_random_idx = 0
            
            # Random and Navigation buttons
            col_random, col_prev, col_next = st.columns(3)
            
            with col_random:
                if st.button("สุ่มรูป", use_container_width=True, key="random_new"):
                    st.session_state.current_random_idx = np.random.randint(0, len(filtered_df))
                    st.rerun()
            
            with col_prev:
                if st.button("ก่อนหน้า", use_container_width=True, key="prev_img"):
                    st.session_state.current_random_idx = (st.session_state.current_random_idx - 1) % len(filtered_df)
                    st.rerun()
            
            with col_next:
                if st.button("ถัดไป", use_container_width=True, key="next_img"):
                    st.session_state.current_random_idx = (st.session_state.current_random_idx + 1) % len(filtered_df)
                    st.rerun()
            
            # Display image info
            selected_idx = st.session_state.current_random_idx
            selected_row = filtered_df.iloc[selected_idx]
            image = Image.open(selected_row["file_path"]).convert("RGB")
            
            col1, col2 = st.columns(2)
            with col1:
                st.image(image, caption=selected_row["file_name"], use_container_width=True)
            with col2:
                st.write(f"**รูป {selected_idx + 1} จาก {len(filtered_df)}**")
                st.success(f"จริง:\n{class_labels.get(selected_row['class_name'])}")
            
            st.divider()
            
            if st.button("ทำนายผล", use_container_width=True, key="predict_btn"):
                pred_class, confidence, probs = predict_image(image, model)
                
                if pred_class:
                    actual_class = selected_row["class_name"]
                    
                    col1, col2 = st.columns(2)
                    with col1:
                        st.info(f"ทำนาย:\n{class_labels.get(pred_class)}")
                        st.metric("ความมั่นใจ", f"{confidence:.1%}")
                    with col2:
                        if pred_class == actual_class:
                            st.success("ถูกต้อง")
                        else:
                            st.error("ผิดพลาด")
                    
                    prob_df = pd.DataFrame({
                        "ประเภท": [class_labels.get(cn, cn) for cn in class_names],
                        "ความน่าจะเป็น": [f"{p:.1%}" for p in probs]
                    }).sort_values("ความน่าจะเป็น", ascending=False)
                    
                    st.dataframe(prob_df, use_container_width=True, hide_index=True)
