requirements.txt
streamlit==1.28.1
pandas>=1.5.0
numpy>=1.23.0
joblib>=1.2.0
matplotlib>=3.6.0
scikit-learn>=1.2.0
pillow>=9.3.0
tensorflow>=2.12.0
keras>=2.12.0
xgboost>=1.7.0
